more template-compiler refactorings
