+++ packages/compat/package.json
@@ -59,7 +59,7 @@
     "semver": "^7.3.2",
     "symlink-or-copy": "^1.2.0",
     "tree-sync": "^2.0.0",
-    "typescript-memoize": "^1.0.0-alpha.3",
+    "typescript-memoize": "^1.0.0",
     "walk-sync": "^1.1.3",
     "yargs": "^16.1.0"
   },
