+++ packages/core/package.json
@@ -51,14 +51,13 @@
     "handlebars": "^4.4.2",
     "js-string-escape": "^1.0.1",
     "jsdom": "^16.4.0",
-    "json-stable-stringify": "^1.0.1",
     "lodash": "^4.17.10",
     "pkg-up": "^3.1.0",
     "resolve": "^1.8.1",
     "resolve-package-path": "^1.2.2",
     "semver": "^7.3.2",
     "strip-bom": "^3.0.0",
-    "typescript-memoize": "^1.0.0-alpha.3",
+    "typescript-memoize": "^1.0.0",
     "walk-sync": "^1.1.3",
     "wrap-legacy-hbs-plugin-if-needed": "^1.0.1"
   },
@@ -69,7 +68,6 @@
     "@types/debug": "^0.0.31",
     "@types/filesize": "^4.1.0",
     "@types/jsdom": "^16.2.4",
-    "@types/json-stable-stringify": "^1.0.32",
     "@types/lodash": "^4.14.116",
     "@types/node": "^10.5.2",
     "@types/resolve": "^0.0.8",
