+++ packages/core/src/app.ts
@@ -19,7 +19,8 @@ import { MacrosConfig } from '@embroider/macros/src/node';
 import { PluginItem, TransformOptions } from '@babel/core';
 import { makePortable } from './portable-babel-config';
 import { TemplateCompilerPlugins } from '.';
-import { templateCompilerModule, TemplateCompilerParams } from './template-compiler';
+import type { TemplateCompilerParams } from './template-compiler';
+import { templateCompilerModule } from './write-template-compiler';
 import { Resolver } from './resolver';
 import { Options as AdjustImportsOptions } from './babel-plugin-adjust-imports';
 import { explicitRelative, extensionsPattern } from './paths';
